// src/wx/typing/jweixin.d.ts
declare namespace wx {
  export type Method =
    'checkJsApi' |
    'updateAppMessageShareData' |
    'updateTimelineShareData' |
    'onMenuShareTimeline' |
    'onMenuShareAppMessage' |
    'onMenuShareQQ' |
    'onMenuShareWeibo' |
    'onMenuShareQZone' |
    'startRecord' |
    'stopRecord' |
    'onVoiceRecordEnd' |
    'playVoice' |
    'pauseVoice' |
    'stopVoice' |
    'onVoicePlayEnd' |
    'uploadVoice' |
    'downloadVoice' |
    'chooseImage' |
    'previewImage' |
    'uploadImage' |
    'downloadImage' |
    'getLocalImgData' |
    'translateVoice' |
    'getNetworkType' |
    'openLocation' |
    'getLocation' |
    'startSearchBeacons' |
    'stopSearchBeacons' |
    'onSearchBeacons' |
    'hideOptionMenu' |
    'showOptionMenu' |
    'hideMenuItems' |
    'showMenuItems' |
    'hideAllNonBaseMenuItem' | // 隐藏所有非基础按钮
    'showAllNonBaseMenuItem' |
    'closeWindow' |
    'scanQRCode' |
    'chooseWXPay' |
    'openProductSpecificView' |
    'openBusinessView' |
    // 'openEnterpriseChat' |
    'addCard' |
    'chooseCard' |
    'openCard';

  export type OpenTag =
    'wx-open-launch-weapp' |
    'wx-open-launch-app' |
    'wx-open-subscribe' |
    'wx-open-audio';
// ...
  interface CommonResponse {
    errMsg: string;
    [k: string]: any;
  }

  interface CommonOptions<
    SuccessResponse extends CommonResponse = CommonResponse,
    FailResponse extends CommonResponse = CommonResponse,
    TriggerResponse extends CommonResponse = CommonResponse
    > {
    success?(res: SuccessResponse): void;
    fail?(res: FailResponse): void;
    complete?(res: SuccessResponse & FailResponse): void;
    cancel?(res: SuccessResponse & FailResponse): void;
    trigger?(res: TriggerResponse): void;
  }

  export interface ConfigOptions {
    /**
     * 公众号的唯一标识
     */
    appId: string;
    /**
     * 生成签名的时间戳
     */
    timestamp: number;
    /**
     * 生成签名的随机串
     */
    nonceStr: string;
    /**
     * 签名
     */
    signature: string;
    /**
     * 需要使用的JS接口列表
     */
    jsApiList?: Method[];
    /**
     * 开启调试模式,调用的所有api的返回值会在客户端alert出来，
     * 若要查看传入的参数，可以在pc端打开，参数信息会通过log打出，仅在pc端时才会打印。
     */
    debug?: boolean;
    /**
     * 需要使用的开放标签列表
     */
    openTagList?: OpenTag[];
  }
// ...
  export interface UpdateAppMessageShareDataOptions extends CommonOptions {
    /** 分享链接，该链接域名或路径必须与当前页面对应的公众号JS安全域名一致 */
    link: string;
    /** 分享标题 */
    title?: string;
    /** 分享描述 */
    desc?: string;
    /** 分享图标 不支持base64*/
    imgUrl?: string;
  }
// ...
}